shuffled monitor into own package
