+++ Dockerfile-build
@@ -15,13 +15,14 @@ ARG VERSION=$VERSION
 ENV VERSION=$VERSION
 
 ### required to build
-RUN mkdir -p /root/butler/internal/stats /root/butler/config /root/butler/config/methods /root/butler/config/reloaders /root/butler/internal/environment /root/butler/internal/alog
+RUN mkdir -p /root/butler/internal/monitor /root/butler/internal/stats /root/butler/config /root/butler/config/methods /root/butler/config/reloaders /root/butler/internal/environment /root/butler/internal/alog
 COPY ./files/build.sh /root/build.sh
 COPY butler.go /root/butler/butler.go
 COPY ./config/*.go /root/butler/config/
 COPY ./config/methods/*.go /root/butler/config/methods/
 COPY ./config/reloaders/*.go /root/butler/config/reloaders/
 COPY ./internal/stats/*.go /root/butler/internal/stats/
+COPY ./internal/monitor/*.go /root/butler/internal/monitor/
 COPY ./internal/environment/*.go /root/butler/internal/environment
 COPY ./internal/alog/*.go /root/butler/internal/alog
 COPY ./vendor /root/butler/vendor
